Allow direct browsing to "nested" datasets sans-manifest
All of the intervening slashes in the dataset path need to be converted
to underscores, not just the first, otherwise the /charon endpoint looks
for file with the wrong name.

This fixes browsing to datasets like /flu/seasonal/h3n2/ha/3y when
running without a manifest.

The same issue was recently fixed for the case when there _is_ a
manifest loaded in commit "Replace all slashes with underscores when
getting dataset path" (952e214).